the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,12 +8,12 @@
  */
 int main (void)
 {
-	long long int n = 50;
-	long long int f = 0, s = 1;
+	long int n = 50;
+	long int f = 0, s = 1;
 	for (int i = 0; i <= n; i++)
 	{
-		printf("%lld, ",f);
-		long long int next = f + s;
+		printf("%ld, ",f);
+		long int next = f + s;
 		f = s;
 		s = next;
 	}
